backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
openpyxl>=3.1.2
pandas>=2.2.0
python-dateutil>=2.8.2
# Google Sheets API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
# Background task scheduling
apscheduler>=3.10.4
